backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "BeaverNotACat", email = "beavernotacat@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.9",
    "sqlalchemy>=2.0.38",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.8.1",
    "alembic[tz]>=1.14.1",
    "psycopg2-binary>=2.9.10",
    "aioboto3>=14.0.0",
    "granian[reload]>=1.7.6",
    "asyncpg>=0.30.0",
    "httpx[http2]>=0.28.1",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "redis>=5.2.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.9.9",
    "mypy>=1.15.0",
    "black>=25.1.0",
    "types-python-dateutil>=2.9.0.20241206",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["I"]

[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true
# TODO: survive with
# strict = true
